Handle delete and backspace on empty lines
When user presses 'backspace' and the caret is in the beginning of a
line and this line is empty this line merges with the previous one.

When user presses 'delete' and the caret is the end of a line, if the next
line is empty, merges with the previous one.

We consider as empty a line without text, so a line with only spaces is empty.