Perform atomic UNDO operation when merging lines with DEL key
Manually process DELETE operation to be able to handle UNDO of line
merging without losing line types (previously we needed to fire UNDO
twice after merging lines with DEL when current line was empty).

This also allowed the tests to trigger the proper DELETE event, so now
all of them are not pending.

The commit also fixes the error that was being raised when user pressed
BACKSPACE at first char of pad, or DEL at last char of pad.