Add classes and scene marks icon in the headings
In this commit we add a scene mark icon in every heading. If there are any scene
mark before this heading, it is showed the icon with the scene mark (A)ct or
(S)equence otherwise the icon is not visible. We mark the scene marks and
heading with special classes as well.