[refactor] Do not use line attrib to format headings for EASC
ep_script_toggle needs some classes on lines with headings, so margins
can be adjusted when when some EASC modes are enabled. Instead of
setting the heading type as a line attribute of each heading, we set it
whenever a heading has its type changed (by adding or removing SMs
to/from it). This simplifies the series of changesets that are generated
everytime a SM is changed.

This is part of the fixes for https://trello.com/c/nvlmyCid/811 and
https://trello.com/c/nvQTxto8/755.